DB에 있는 카페를 필터링 하는 기능 추가. 카카오 API랑 조건을 붙여 합치기 전 임시저장 커밋.

+++ src/main/java/com/skkucapstone/Castardbackend/domain/Cafe.java
@@ -21,9 +21,9 @@ public class Cafe {
 
     private String phone;
 
-    private Long longitude; // x좌표
+    private Double longitude; // x좌표
 
-    private Long latitude; // y좌표
+    private Double latitude; // y좌표
 
     // 카공에 대한 리뷰들의 평균을 유지하는 필드들.
     private Double power_socket = 0D;
